% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_buttons.R
\name{button_send_datamap}
\alias{button_send_datamap}
\title{button_send_datamap}
\usage{
button_send_datamap(rv)
}
\arguments{
\item{rv}{The global rv object. rv$datamap and rv$config_file need to
be valid.}
}
\description{
This function is an exporte wrapper around the actual function
  to send the datamap. This actual function can be customized by the user.
}
