% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_buttons.R
\name{get_influx_connection}
\alias{get_influx_connection}
\title{get_influx_connection}
\usage{
get_influx_connection(rv)
}
\arguments{
\item{rv}{The global rv object. Only the rv$config_file part will be used.}
}
\value{
Returns a list with elements result$con (The connection to influxdb)
  and result$config (The config credentials extracted from the rv-object).
}
\description{
This function loads the datamap from the rv-object and sends
  it to an influxdb server. The credentials will be loaded from the
  config_file also stored in the rv-object.
  If the export was successfull, rv$datamap$exported will be set to TRUE.
}
