% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchDQAtool}
\alias{launchDQAtool}
\title{Launch the MIRACUM DQA Tool}
\usage{
launchDQAtool(port = 3838)
}
\arguments{
\item{port}{The port, the MIRACUM DQA Tool is running on (default: 3838)}
}
\value{
the MIRACUM DQA Tool Shiny application
}
\description{
Launch the MIRACUM DQA Tool
}
