% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_from_samply.R
\name{mdr_from_samply}
\alias{mdr_from_samply}
\title{DQA-MDR to Samply.MDR Converter Function}
\usage{
mdr_from_samply(
  base_url = "https://mdr.miracum.de/rest/api/mdr/",
  namespace = "dqa",
  master_system_type = "csv",
  master_system_name = "p21csv",
  headless = TRUE,
  logfile_dir
)
}
\arguments{
\item{base_url}{The URL of the Samply.MDR, from which the `namespace`
should be imported.}

\item{namespace}{A character string. The name of the DQA-MDR namespace in
the Samply.MDR (default: 'dqa').}

\item{master_system_type}{A character string. The type of the master system
(the master system contains more information on the data elements;
currently, 'csv' is the only allowed value.)}

\item{master_system_name}{A character string. The name of the master system
(the master system contains more information on the data elements;
currently, 'p21csv' is the only allowed value.)}

\item{headless}{A boolean (default: TRUE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}

\item{logfile_dir}{A character string. Path to the logfile.}
}
\description{
DQA-MDR to Samply.MDR Converter Function
}
\references{
\cite{D. Kadioglu, B. Breil, C. Knell, M. Lablans, S. Mate,
  D. Schlue, H. Serve, H. Storf, F. Ückert, T. Wagner, P. Weingardt,
  and H.-U. Prokosch, Samply.MDR - A Metadata Repository and Its
  Application in Various Research Networks, Studies in Health Technology
  and Informatics. (2018) 50–54. doi:10/gf963s.}
}
\seealso{
\url{http://ebooks.iospress.nl/publication/50022}
}
