% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_to_samply.R
\name{mdr_to_samply}
\alias{mdr_to_samply}
\title{DQA-MDR to Samply.MDR Converter Function}
\usage{
mdr_to_samply(
  utilspath = "inst/application/_utilities/",
  mdr_filename = "mdr.csv",
  master_system_name = "p21csv"
)
}
\arguments{
\item{utilspath}{The path to the utilities-folder, containing the metadata
repository files (`mdr.csv` inside the folder `MDR`), JSON files with
SQL statements (inside the folder `SQL`), config files for the database
connection (`settings_default.yml`) and the email address used for the
data map (`email.yml`), a JSON file containing site names (inside the
folder `MISC`) and a markdown templated to create the PDF report
(`DQA_report.Rmd` inside the folder `RMD`).}

\item{mdr_filename}{A character string. The name of the DQA-MDR CSV file
(default: 'mdr.csv').}

\item{master_system_name}{The name of the master system. Currently, the only
allowed argument is `p21csv`. All necessary MDR items, including
`definition` and `dqa_assessment`, are stored on the master system level
in the metadata repository.}
}
\description{
DQA-MDR to Samply.MDR Converter Function
}
