% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{miRacumDQA}
\alias{miRacumDQA}
\title{Launch the MIRACUM DQA Tool}
\usage{
miRacumDQA(port = 3838)
}
\arguments{
\item{port}{The port, miRacumDQA is running on (default: 3838)}
}
\value{
miRacum DQA Shiny application
}
\description{
Launch the MIRACUM DQA Tool
}
