% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAtempPlausibility.R
\name{moduleAtempPlausibilityUI}
\alias{moduleAtempPlausibilityUI}
\title{moduleAtempPlausibilityUI}
\usage{
moduleAtempPlausibilityUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleAtempPlausibilityUI
}
