% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDashboard.R
\name{moduleDashboardUI}
\alias{moduleDashboardUI}
\title{moduleDashboardUI}
\usage{
moduleDashboardUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleDashboardUI
}
