% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDescriptive.R
\name{moduleDescriptiveServer}
\alias{moduleDescriptiveServer}
\title{moduleDescriptiveServer}
\usage{
moduleDescriptiveServer(input, output, session, rv, input_re)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{input_re}{The Shiny server input object, wrapped into a reactive expression: input_re = reactive({input})}
}
\description{
moduleDescriptiveServer
}
