% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDescriptive.R
\name{moduleDescriptiveUI}
\alias{moduleDescriptiveUI}
\title{moduleDescriptiveUI}
\usage{
moduleDescriptiveUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleDescriptiveUI
}
