% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleRawdata1.R
\name{moduleRawdata1UI}
\alias{moduleRawdata1UI}
\title{moduleRawdata1UI}
\usage{
moduleRawdata1UI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleRawdata1UI
}
