% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleUniquePlausibility.R
\name{moduleUniquePlausibilityUI}
\alias{moduleUniquePlausibilityUI}
\title{moduleUniquePlausibilityUI}
\usage{
moduleUniquePlausibilityUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleUniquePlausibilityUI
}
