% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleVisualizations.R
\name{moduleVisualizationsUI}
\alias{moduleVisualizationsUI}
\title{moduleVisualizationsUI}
\usage{
moduleVisualizationsUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleVisualizationsUI
}
