% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_datamap_to_influxdb.R
\name{send_datamap_to_api}
\alias{send_datamap_to_api}
\title{send_datamap_to_api}
\usage{
send_datamap_to_api(rv)
}
\arguments{
\item{rv}{The global rv object. Only the rv$datamap part will be used.}
}
\description{
send_datamap_to_api
}
