% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_buttons.R
\name{send_datamap_to_influx}
\alias{send_datamap_to_influx}
\title{send_datamap_to_influx}
\usage{
send_datamap_to_influx(rv)
}
\arguments{
\item{rv}{The global rv object. rv$datamap and rv$config_file need to
be valid.}
}
\description{
This function loads the datamap from the rv-object and sends
  it to an influxdb server. The credentials will be loaded from the
  config_file also stored in the rv-object.
  If the export was successfull, rv$datamap$exported will be set to TRUE.
}
