% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_dqa_tool}
\alias{launch_dqa_tool}
\title{Launch the MIRACUM DQA Tool}
\usage{
launch_dqa_tool(
  port = 3838,
  utils_path = system.file("application/_utilities", package = "miRacumDQA"),
  logfile_dir = tempdir()
)
}
\arguments{
\item{port}{The port, the MIRACUM DQA Tool is running on (default: 3838)}

\item{utils_path}{The path to the utilities-folder, containing the metadata
repository files (`mdr.csv` inside the folder `MDR`), JSON files with SQL
statements (inside the folder `SQL`), config files for the database
connection (`settings_default.yml`) and the email address used for the
data map (`email.yml`), a JSON file containing site names (inside the
folder `MISC`) and a markdown templated to create the PDF report
(`DQA_report.Rmd` inside the folder `RMD`).}

\item{logfile_dir}{Is the absolute path to the directory where the logfile
will be stored. If not path is provided the tempdir() will be used.}
}
\value{
the MIRACUM DQA Tool Shiny application
}
\description{
Launch the MIRACUM DQA Tool
}
