% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchDQAtool}
\alias{launchDQAtool}
\title{Launch the MIRACUM DQA Tool}
\usage{
launchDQAtool(port = 3838, utilspath, db_source)
}
\arguments{
\item{port}{The port, the MIRACUM DQA Tool is running on (default: 3838)}

\item{utilspath}{The path to the utilities-folder, containing the metadata repository files (`mdr.csv` inside the folder `MDR`),
JSON files with SQL statements (inside the folder `SQL`), config files for the database connection (`settings_default.yml`)
and the email address used for the data map (`email.yml`), a JSON file containing site names (inside the folder `MISC`) and a
markdown templated to create the PDF report (`DQA_report.Rmd` inside the folder `RMD`).}

\item{db_source}{The name of the source database. Currently, the only allowed argument is `p21csv` used for the import of the
corresponding csv files.}
}
\value{
the MIRACUM DQA Tool Shiny application
}
\description{
Launch the MIRACUM DQA Tool
}
